Move bor_receipts to polygon
