+++ rpc/jsonrpc/eth_txs.go
@@ -30,6 +30,7 @@ import (
 	types2 "github.com/erigontech/erigon-lib/types"
 	"github.com/erigontech/erigon/erigon-db/rawdb"
 	bortypes "github.com/erigontech/erigon/polygon/bor/types"
+	borrawdb "github.com/erigontech/erigon/polygon/rawdb"
 	"github.com/erigontech/erigon/rpc"
 	"github.com/erigontech/erigon/rpc/ethapi"
 	"github.com/erigontech/erigon/rpc/rpchelper"
@@ -227,7 +228,7 @@ func (api *APIImpl) GetTransactionByBlockHashAndIndex(ctx context.Context, block
 				borTx = bortypes.NewBorTransaction()
 			}
 		} else {
-			borTx = rawdb.ReadBorTransactionForBlock(tx, block.NumberU64())
+			borTx = borrawdb.ReadBorTransactionForBlock(tx, block.NumberU64())
 		}
 		if borTx == nil {
 			return nil, nil // not error
@@ -303,7 +304,7 @@ func (api *APIImpl) GetTransactionByBlockNumberAndIndex(ctx context.Context, blo
 				borTx = bortypes.NewBorTransaction()
 			}
 		} else {
-			borTx = rawdb.ReadBorTransactionForBlock(tx, blockNum)
+			borTx = borrawdb.ReadBorTransactionForBlock(tx, blockNum)
 		}
 		if borTx == nil {
 			return nil, nil
